% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{begin}
\alias{begin}
\title{Begin Exam}
\usage{
begin(sisi_id, password, exam_id)
}
\arguments{
\item{sisi_id}{string, the SISI ID of the student}

\item{password}{string, the exam password which is assigned to the student from \href{https://www.magadlal.com/}{www.magadlal.com}}

\item{exam_id}{integer, exam ID on \href{https://www.magadlal.com/}{www.magadlal.com}}
}
\value{
\code{NULL}. It prints a response message from the server.
}
\description{
It begins an online exam session on \href{https://www.magadlal.com/}{www.magadlal.com}.
}
\seealso{
\code{\link{problem}}, \code{\link{check}}
}
