% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{begin_exam}
\alias{begin_exam}
\title{Begin Exam}
\usage{
begin_exam(sisi_id, password, exam_id)
}
\arguments{
\item{sisi_id}{string, the SISi ID of the student}

\item{password}{string, the exam password which is assigned to the student from \href{https://www.magadlal.com/}{www.magadlal.com}}

\item{exam_id}{string, exam ID on \href{https://www.magadlal.com/}{www.magadlal.com}}
}
\value{
\code{NULL}. It prints a response message from the server.
}
\description{
It begins a new online exam session on \href{https://www.magadlal.com/}{www.magadlal.com}.
}
