% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{check}
\alias{check}
\title{Check Problem Solution}
\usage{
check(sisi_id, exam_id, expr)
}
\arguments{
\item{sisi_id}{string, the SISI ID of the student}

\item{exam_id}{integer, exam ID on \href{https://www.magadlal.com/}{www.magadlal.com}}

\item{expr}{an expression, an R code which is a solution}
}
\value{
Typically, either \code{TRUE} or a string describing the differences between students answer and the correct answer of the assigned problem from the server. When an error occurs, \code{NULL} is returned.
}
\description{
This function checks a solution and submits it to the server when the solution is correct.
}
\seealso{
\code{\link{begin}}, \code{\link{problem}}
}
