% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{check_answer}
\alias{check_answer}
\title{Check Answer}
\usage{
check_answer(answer)
}
\arguments{
\item{answer}{an object or expression, an answer from a student}
}
\value{
Either \code{TRUE} or a string describing the differences between students answer and the correct answer of the assigned problem from the server.
}
\description{
It checks an answer.
}
