% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{get_problem}
\alias{get_problem}
\title{Get Problem}
\usage{
get_problem(prevent_new_problem = FALSE)
}
\arguments{
\item{prevent_new_problem}{If \code{TRUE}, it prevents from getting a new problem accidentally and attempts to get the problem which is currently assigned.}
}
\value{
A list which is contains a new or current problem is assigned to a student from the server.
}
\description{
Get a new or current problem which is assigned to a student from the server.
}
