% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{is.completed}
\alias{is.completed}
\title{Check Exam is Completed}
\usage{
is.completed()
}
\value{
\code{NULL}. It prints a response message from the server.
}
\description{
I checks whether an exam is completed or not.
}
