% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{problem}
\alias{problem}
\title{Get Problem}
\usage{
problem(sisi_id, exam_id, prevent_new_problem = FALSE)
}
\arguments{
\item{sisi_id}{string, the SISI ID of the student}

\item{exam_id}{integer, exam ID on \href{https://www.magadlal.com/}{www.magadlal.com}}

\item{prevent_new_problem}{If \code{TRUE}, it prevents from getting a new problem accidentally and attempts to get the problem which is currently assigned.}
}
\value{
A list which is contains a new or current problem is assigned to a student from the server.
}
\description{
Get a new or current problem which is assigned to a student from the server.
}
\seealso{
\code{\link{begin}}, \code{\link{check}}
}
