% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{remaining_time}
\alias{remaining_time}
\title{Remaining Time}
\usage{
remaining_time()
}
\value{
\code{NULL}. It prints remaining time for the current problem.
}
\description{
Check remaining time from the server for the current problem.
}
