% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{skip_problem}
\alias{skip_problem}
\title{Skip current problem}
\usage{
skip_problem()
}
\value{
\code{NULL}. It prints a response message from the server.
}
\description{
It skips a problem which is currently assigned to a student from the server.
}
