% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{submit_solution}
\alias{submit_solution}
\title{Check and Submit a Solution}
\usage{
submit_solution(uid, expr)
}
\arguments{
\item{uid}{integer, the unique identification number that is assigned to the student via the exam page of the \href{https://www.magadlal.com/}{www.magadlal.com}}

\item{expr}{an expression, an R code which is a solution}
}
\value{
Typically, a server response message or an error message, if a student's answer has an error.
}
\description{
This function checks a solution and submits it to the server when it is correct.
}
