% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{begin_exam}
\alias{begin_exam}
\title{Begin Exam}
\usage{
begin_exam(sisi_id, password, exam_id)
}
\arguments{
\item{sisi_id}{string, the SISI ID of the student}

\item{password}{string, the exam password that is assigned to the student from \href{https://www.magadlal.com/}{www.magadlal.com}}

\item{exam_id}{integer, exam ID on \href{https://www.magadlal.com/}{www.magadlal.com}}
}
\value{
It returns a server response message.
}
\description{
It begins an online exam session on \href{https://www.magadlal.com/}{www.magadlal.com}.
}
\seealso{
\code{\link{get_problem}}, \code{\link{check_solution}}
}
