% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{get_problem}
\alias{get_problem}
\title{Get Problem}
\usage{
get_problem(sisi_id, exam_id, prevent_new_problem = FALSE)
}
\arguments{
\item{sisi_id}{string, a SISI ID of a student}

\item{exam_id}{integer, exam ID on \href{https://www.magadlal.com/}{www.magadlal.com}}

\item{prevent_new_problem}{If \code{TRUE}, it prevents from getting a new problem accidentally and attempts to get the problem which is assigned recently.}
}
\value{
A list which contains a new or current problem is assigned to a student by the server.
}
\description{
Get a new or current problem is assigned to a student by the server.
}
\seealso{
\code{\link{begin_exam}}, \code{\link{check_solution}}
}
